MCR-2987 Moved resolver to extra class
