+++ mycore-base/src/main/java/org/mycore/common/xml/MCRURIResolver.java
@@ -23,7 +23,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.PrintWriter;
 import java.io.StringWriter;
-import java.lang.reflect.Method;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
@@ -248,7 +247,6 @@ public final class MCRURIResolver implements URIResolver {
         supportedSchemes.put("https", restResolver);
         supportedSchemes.put("file", new MCRFileResolver());
         supportedSchemes.put("cache", new MCRCachingResolver());
-        supportedSchemes.put("call-java", new MCRFunctionResolver());
         return supportedSchemes;
     }
 
@@ -1826,47 +1824,6 @@ public final class MCRURIResolver implements URIResolver {
             }
 
             return new JDOMSource(resolvedXML);
-        }
-    }
-
-    /**
-     * Resolves arbitrary static methods of arbitrary classes. Parameters are considerd to be of type
-     * {@link java.lang.String}.
-     * <br/><br/>
-     * <strong>Invocation</strong>
-     * <pre><code>function:&lt;class name&gt;:&lt;method name&gt;:&lt;param1&gt;:&lt;param2&gt;</code></pre>
-     * <br/>
-     * <strong>Example</strong>
-     * <pre><code>function:de.uni_jena.thunibib.user.ThUniBibUtils:getLeadId:id_connection:foobar;</code></pre>
-     *
-     * @author shermann (Silvio Hermann)
-     * */
-    private static class MCRFunctionResolver implements URIResolver {
-        @Override
-        public Source resolve(String href, String base) throws TransformerException {
-            LOGGER.debug("Resolving {}", href);
-
-            String[] parts = href.split(":");
-            String className = parts[1];
-            String methodName = parts[2];
-
-            Object[] params = Arrays.stream(parts).skip(3).toArray(String[]::new);
-
-            try {
-                Class[] types = new Class[params.length];
-                Arrays.fill(types, String.class);
-
-                Object result = null;
-                Method method = MCRClassTools.forName(className).getMethod(methodName, types);
-                result = method.invoke(null, params);
-
-                Element string = new Element("string");
-                string.setText(result == null ? "" : String.valueOf(result));
-
-                return new JDOMSource(string);
-            } catch (Exception e) {
-                throw new TransformerException(e);
-            }
         }
     }
 }
