+++ mycore-base/src/main/resources/config/mycore.properties
@@ -67,6 +67,7 @@ MCR.Filter.UserAgent.AcceptInvalid=false
 
   MCR.URIResolver.ModuleResolver.property=org.mycore.common.xml.MCRPropertiesResolver
   MCR.URIResolver.ModuleResolver.userobjectrights=org.mycore.common.xml.MCRUserAndObjectRightsURIResolver
+  MCR.URIResolver.ModuleResolver.callJava=org.mycore.common.xml.MCRFunctionResolver
 
 # Caching URIResolver: cache:...
